Zork 4.2
Changed the method of file reading from a .txt file to a JSON file. Downloaded the Newtonsoft NuGet package to get access to JSON classes. Used deserialization to read from the JSON file and initialize all the rooms with their names and descriptions into the 2D array of rooms. The rooms are no longer hard coded in the source code.